refactor: 폴더 경로 수정 및 타입 분리
