+++ src/app/(route)/_components/Onboarding/_constants/onBoardingAnimations.ts
