+++ src/app/(route)/_components/Onboarding/_type/type.ts
@@ -1,0 +1,6 @@
+export interface OnboardingStep {
+  id: number;
+  title: string;
+  description: string;
+  imageSrc: string;
+}
